Add support for M1 macs
This commit allows any users who are on M1 to use the plugin.
This is achieved by using my fork of the DiscordIPC library insted of Minnced's DiscordRPC library. The DiscordIPC library should work across all platforms.

The only known issue is that we are unable to get the user which is logged in to discord to display in the settings tab.