Use IPC if on an ARM CPU, otherwise use RPC
This commit restores the original functionality of NativeRpcConnection
and moves the IPC code to IPCConnection.kt.

This allows RPC to be used where we know that it is available, IPC will
be used if the architecture is aarch64.

You can force IPC / RPC by modifying an environment variable:
`com.almightyalpaca.jetbrains.plugins.discord.plugin.rpc.connection`.
Setting this to "rpc" or "ipc" will change the connection type used.